Optimizations

+++ src/header/maputils.h
@@ -4,7 +4,7 @@
 #include <im.h>
 #include <im_image.h>
 
-int nearest_color_index(unsigned char, unsigned char, unsigned char);
+short nearest_color_index(unsigned char, unsigned char, unsigned char);
 imImage* mapify(imImage*);
 
 #endif /* SRC_HEADER_MAPUTILS_H_ */
